Refactor dirs
