+++ src/error/index.ts
@@ -1,0 +1,2 @@
+export * from "./panic"
+export * from "./result_error"
