+++ src/index.ts
@@ -1,10 +1,5 @@
-export * from "./fn"
-export * from "./guard"
+export * from "./error"
+export * from "./helpers"
 export * from "./option"
-export * from "./panic"
-export * from "./promise_option"
-export * from "./promise_result"
-export * from "./result_error"
 export * from "./result"
-export * from "./try"
 export * from "./util"
