+++ src/option/index.ts
@@ -1,0 +1,4 @@
+export * from "./interface"
+export * from "./none"
+export * from "./some"
+export * from "./promise"
