+++ src/option/promise_option.test.ts
@@ -1,7 +1,5 @@
 import {describe, it, expect, vi} from "vitest"
-import {Panic, UnwrapPanic} from "../src"
-import {PromiseOption} from "../src/promise_option"
-import {None, Some} from "../src/option"
+import {Panic, UnwrapPanic, PromiseOption, Some, None} from ".."
 
 function promiseSome<T>(value: T) {
 	return new PromiseOption(Promise.resolve(Some<T>(value)))
