+++ src/result/index.ts
@@ -1,0 +1,4 @@
+export * from "./err"
+export * from "./interface"
+export * from "./ok"
+export * from "./promise"
