+++ src/result/promise_result.test.ts
@@ -1,5 +1,5 @@
 import {describe, it, expect} from "vitest"
-import {Err, Panic, PromiseResult, Ok, UnwrapPanic} from "../src"
+import {Err, Panic, PromiseResult, Ok, UnwrapPanic} from ".."
 
 function promiseOk<T, E = any>(value: T) {
 	return new PromiseResult<T, E>(Promise.resolve(Ok<T>(value)))
